+++ crates/moon/tests/test_cases/fmt/lib/hello.mbt
