+++ crates/moon/tests/test_cases/fmt/lib/moon.pkg.json
