+++ crates/moon/tests/test_cases/fmt/main/main.mbt
