+++ crates/moon/tests/test_cases/fmt/main/moon.pkg.json
