+++ crates/moon/tests/test_cases/mod.rs
@@ -37,6 +37,7 @@ mod docs_examples;
 mod dummy_core;
 mod extra_flags;
 mod fancy_import;
+mod fmt;
 mod hello;
 mod inline_test;
 mod mbti;
@@ -54,6 +55,7 @@ mod targets;
 mod test_error_report;
 mod test_expect_test;
 mod test_filter;
+mod warns;
 
 #[test]
 fn test_need_link() {
@@ -65,200 +67,6 @@ fn test_need_link() {
             moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg username/hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources username/hello/main:./main -target wasm-gc
             moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main username/hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources username/hello/lib:./lib -pkg-sources username/hello/main:./main -target wasm-gc
             moonc link-core ./target/wasm-gc/release/build/lib/lib.core -main username/hello/lib -o ./target/wasm-gc/release/build/lib/lib.wasm -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -target wasm-gc
-        "#]],
-    );
-}
-
-#[test]
-fn test_warn_list_dry_run() {
-    let dir = TestDir::new("warn_list.in");
-
-    check(
-        get_stdout(&dir, ["build", "--sort-input", "--no-render", "--dry-run"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -w -2 -o ./target/wasm-gc/release/build/lib/lib.core -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:./lib -target wasm-gc
-            moonc build-package ./lib1/hello.mbt -w -1 -o ./target/wasm-gc/release/build/lib1/lib1.core -pkg username/hello/lib1 -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib1:./lib1 -target wasm-gc
-            moonc build-package ./main/main.mbt -w -1-2 -o ./target/wasm-gc/release/build/main/main.core -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/build/lib/lib.mi:lib -i ./target/wasm-gc/release/build/lib1/lib1.mi:lib1 -pkg-sources username/hello/main:./main -target wasm-gc
-            moonc link-core $MOON_HOME/lib/core/target/wasm-gc/release/bundle/core.core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/lib1/lib1.core ./target/wasm-gc/release/build/main/main.core -main username/hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources username/hello/lib:./lib -pkg-sources username/hello/lib1:./lib1 -pkg-sources username/hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            [
-                "build",
-                "--warn-list",
-                "-29",
-                "--sort-input",
-                "--no-render",
-                "--dry-run",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -w -2-29 -o ./target/wasm-gc/release/build/lib/lib.core -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:./lib -target wasm-gc
-            moonc build-package ./lib1/hello.mbt -w -1-29 -o ./target/wasm-gc/release/build/lib1/lib1.core -pkg username/hello/lib1 -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib1:./lib1 -target wasm-gc
-            moonc build-package ./main/main.mbt -w -1-2-29 -o ./target/wasm-gc/release/build/main/main.core -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/build/lib/lib.mi:lib -i ./target/wasm-gc/release/build/lib1/lib1.mi:lib1 -pkg-sources username/hello/main:./main -target wasm-gc
-            moonc link-core $MOON_HOME/lib/core/target/wasm-gc/release/bundle/core.core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/lib1/lib1.core ./target/wasm-gc/release/build/main/main.core -main username/hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources username/hello/lib:./lib -pkg-sources username/hello/lib1:./lib1 -pkg-sources username/hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target wasm-gc
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["test", "--sort-input"]),
-        expect![[r#"
-            Total tests: 1, passed: 1, failed: 0.
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["bundle", "--sort-input", "--no-render", "--dry-run"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -w -2 -o ./target/wasm-gc/release/bundle/lib/lib.core -pkg username/hello/lib -pkg-sources username/hello/lib:./lib -target wasm-gc
-            moonc build-package ./lib1/hello.mbt -w -1 -o ./target/wasm-gc/release/bundle/lib1/lib1.core -pkg username/hello/lib1 -pkg-sources username/hello/lib1:./lib1 -target wasm-gc
-            moonc build-package ./main/main.mbt -w -1-2 -o ./target/wasm-gc/release/bundle/main/main.core -pkg username/hello/main -is-main -i ./target/wasm-gc/release/bundle/lib/lib.mi:lib -i ./target/wasm-gc/release/bundle/lib1/lib1.mi:lib1 -pkg-sources username/hello/main:./main -target wasm-gc
-            moonc bundle-core ./target/wasm-gc/release/bundle/lib/lib.core ./target/wasm-gc/release/bundle/lib1/lib1.core ./target/wasm-gc/release/bundle/main/main.core -o ./target/wasm-gc/release/bundle/hello.core
-        "#]],
-    );
-
-    // to cover `moon bundle` no work to do
-    get_stdout(&dir, ["bundle", "--sort-input"]);
-
-    check(
-        get_stdout(&dir, ["check", "--sort-input", "--no-render", "--dry-run"]),
-        expect![[r#"
-            moonc check ./lib/hello.mbt -w -2 -o ./target/wasm-gc/release/check/lib/lib.mi -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:./lib -target wasm-gc
-            moonc check ./lib1/hello.mbt -w -1 -o ./target/wasm-gc/release/check/lib1/lib1.mi -pkg username/hello/lib1 -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib1:./lib1 -target wasm-gc
-            moonc check ./main/main.mbt -w -1-2 -o ./target/wasm-gc/release/check/main/main.mi -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/check/lib/lib.mi:lib -i ./target/wasm-gc/release/check/lib1/lib1.mi:lib1 -pkg-sources username/hello/main:./main -target wasm-gc
-            moonc check ./lib/hello_test.mbt -w -2 -o ./target/wasm-gc/release/check/lib/lib.blackbox_test.mi -pkg username/hello/lib_blackbox_test -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/check/lib/lib.mi:lib -pkg-sources username/hello/lib_blackbox_test:./lib -target wasm-gc -blackbox-test
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            [
-                "check",
-                "--warn-list",
-                "-29",
-                "--sort-input",
-                "--no-render",
-                "--dry-run",
-            ],
-        ),
-        expect![[r#"
-            moonc check ./lib/hello.mbt -w -2-29 -o ./target/wasm-gc/release/check/lib/lib.mi -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:./lib -target wasm-gc
-            moonc check ./lib1/hello.mbt -w -1-29 -o ./target/wasm-gc/release/check/lib1/lib1.mi -pkg username/hello/lib1 -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib1:./lib1 -target wasm-gc
-            moonc check ./main/main.mbt -w -1-2-29 -o ./target/wasm-gc/release/check/main/main.mi -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/check/lib/lib.mi:lib -i ./target/wasm-gc/release/check/lib1/lib1.mi:lib1 -pkg-sources username/hello/main:./main -target wasm-gc
-            moonc check ./lib/hello_test.mbt -w -2-29 -o ./target/wasm-gc/release/check/lib/lib.blackbox_test.mi -pkg username/hello/lib_blackbox_test -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/check/lib/lib.mi:lib -pkg-sources username/hello/lib_blackbox_test:./lib -target wasm-gc -blackbox-test
-        "#]],
-    );
-}
-
-#[test]
-fn test_warn_list_real_run() {
-    let dir = TestDir::new("warn_list.in");
-
-    check(
-        get_stderr(&dir, ["build", "--sort-input", "--no-render"]),
-        expect![[r#"
-            Finished. moon: ran 4 tasks, now up to date
-        "#]],
-    );
-
-    check(
-        get_stderr(&dir, ["test", "--sort-input"])
-            .lines()
-            .filter(|it| !it.starts_with("Blocking waiting for file lock"))
-            .collect::<String>(),
-        expect![[r#""#]],
-    );
-    check(
-        get_stdout(&dir, ["test", "--sort-input"]),
-        expect![[r#"
-            Total tests: 1, passed: 1, failed: 0.
-        "#]],
-    );
-
-    check(
-        get_stderr(&dir, ["bundle", "--sort-input", "--no-render"]),
-        expect![[r#"
-            Finished. moon: ran 4 tasks, now up to date
-        "#]],
-    );
-
-    // to cover `moon bundle` no work to do
-    get_stdout(&dir, ["bundle", "--sort-input"]);
-
-    check(
-        get_stderr(&dir, ["check", "--sort-input", "--no-render"]),
-        expect![[r#"
-            Finished. moon: ran 4 tasks, now up to date
-        "#]],
-    );
-}
-
-#[test]
-fn test_alert_list() {
-    std::env::set_var("NO_COLOR", "1");
-    let dir = TestDir::new("alert_list.in");
-
-    check(
-        get_stderr(&dir, ["build", "--sort-input"]),
-        expect![[r#"
-            Warning: [2000]
-               ╭─[$ROOT/main/main.mbt:3:3]
-               │
-             3 │   alert_2();
-               │   ───┬───  
-               │      ╰───── Warning (Alert alert_2): alert_2
-            ───╯
-            Finished. moon: ran 3 tasks, now up to date
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["test", "--sort-input"]),
-        expect![[r#"
-            Total tests: 1, passed: 1, failed: 0.
-        "#]],
-    );
-
-    check(
-        get_stderr(&dir, ["bundle", "--sort-input"]),
-        expect![[r#"
-            Warning: [2000]
-               ╭─[$ROOT/main/main.mbt:3:3]
-               │
-             3 │   alert_2();
-               │   ───┬───  
-               │      ╰───── Warning (Alert alert_2): alert_2
-            ───╯
-            Finished. moon: ran 3 tasks, now up to date
-        "#]],
-    );
-
-    check(
-        get_stderr(&dir, ["check", "--sort-input"]),
-        expect![[r#"
-            Warning: [2000]
-               ╭─[$ROOT/main/main.mbt:3:3]
-               │
-             3 │   alert_2();
-               │   ───┬───  
-               │      ╰───── Warning (Alert alert_2): alert_2
-            ───╯
-            Finished. moon: ran 2 tasks, now up to date
-        "#]],
-    );
-}
-
-#[test]
-fn test_mod_level_warn_alert_list() {
-    let dir = TestDir::new("mod_level_warn&alert_list.in");
-
-    check(
-        get_stdout(&dir, ["check", "--dry-run"]),
-        expect![[r#"
-            moonc check ./lib/hello.mbt -w -1 -alert -alert_1 -o ./target/wasm-gc/release/check/lib/lib.mi -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:./lib -target wasm-gc
-            moonc check ./main/main.mbt -w -1-2 -alert -alert_1-alert_2 -o ./target/wasm-gc/release/check/main/main.mi -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -i ./target/wasm-gc/release/check/lib/lib.mi:lib -pkg-sources username/hello/main:./main -target wasm-gc
         "#]],
     );
 }
@@ -331,319 +139,6 @@ fn test_backtrace() {
     let out = get_err_stderr(&dir, ["run", "main", "--debug"]);
     assert!(out.contains("main.foo"));
     assert!(out.contains("main.bar"));
-}
-
-#[test]
-fn test_deny_warn() {
-    std::env::set_var("NO_COLOR", "1");
-    let dir = TestDir::new("test_deny_warn.in");
-
-    check(
-        get_stderr(&dir, ["check", "--sort-input"]),
-        expect![[r#"
-            Warning: [2000]
-                ╭─[$ROOT/lib/hello.mbt:13:3]
-                │
-             13 │   alert_1();
-                │   ───┬───  
-                │      ╰───── Warning (Alert alert_1): alert_1
-            ────╯
-            Warning: [2000]
-                ╭─[$ROOT/lib/hello.mbt:14:3]
-                │
-             14 │   alert_2();
-                │   ───┬───  
-                │      ╰───── Warning (Alert alert_2): alert_2
-            ────╯
-            Warning: [1002]
-               ╭─[$ROOT/lib/hello.mbt:4:7]
-               │
-             4 │   let a = 1;
-               │       ┬  
-               │       ╰── Warning: Unused variable 'a'
-            ───╯
-            Warning: [1002]
-                ╭─[$ROOT/lib/hello.mbt:11:7]
-                │
-             11 │   let 中文 = 2
-                │       ──┬─  
-                │         ╰─── Warning: Unused variable '中文'
-            ────╯
-            Warning: [1002]
-                ╭─[$ROOT/lib/hello.mbt:12:7]
-                │
-             12 │   let 🤣😭🤣😭🤣 = 2
-                │       ────┬─────  
-                │           ╰─────── Warning: Unused variable '🤣😭🤣😭🤣'
-            ────╯
-            Warning: [1002]
-               ╭─[$ROOT/main/main.mbt:2:7]
-               │
-             2 │   let a = 0
-               │       ┬  
-               │       ╰── Warning: Unused variable 'a'
-            ───╯
-            Finished. moon: ran 2 tasks, now up to date
-        "#]],
-    );
-
-    check(
-        get_err_stdout(&dir, ["check", "--deny-warn", "--sort-input"]),
-        expect![[r#"
-            failed: moonc check -error-format json -w @a-31-32 -alert @all-raise-throw-unsafe+deprecated $ROOT/lib/hello.mbt -o $ROOT/target/wasm-gc/release/check/lib/lib.mi -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:$ROOT/lib -target wasm-gc
-        "#]],
-    );
-
-    check(
-        get_stderr(&dir, ["build", "--sort-input"]),
-        expect![[r#"
-            Warning: [2000]
-                ╭─[$ROOT/lib/hello.mbt:13:3]
-                │
-             13 │   alert_1();
-                │   ───┬───  
-                │      ╰───── Warning (Alert alert_1): alert_1
-            ────╯
-            Warning: [2000]
-                ╭─[$ROOT/lib/hello.mbt:14:3]
-                │
-             14 │   alert_2();
-                │   ───┬───  
-                │      ╰───── Warning (Alert alert_2): alert_2
-            ────╯
-            Warning: [1002]
-               ╭─[$ROOT/lib/hello.mbt:4:7]
-               │
-             4 │   let a = 1;
-               │       ┬  
-               │       ╰── Warning: Unused variable 'a'
-            ───╯
-            Warning: [1002]
-                ╭─[$ROOT/lib/hello.mbt:11:7]
-                │
-             11 │   let 中文 = 2
-                │       ──┬─  
-                │         ╰─── Warning: Unused variable '中文'
-            ────╯
-            Warning: [1002]
-                ╭─[$ROOT/lib/hello.mbt:12:7]
-                │
-             12 │   let 🤣😭🤣😭🤣 = 2
-                │       ────┬─────  
-                │           ╰─────── Warning: Unused variable '🤣😭🤣😭🤣'
-            ────╯
-            Warning: [1002]
-               ╭─[$ROOT/main/main.mbt:2:7]
-               │
-             2 │   let a = 0
-               │       ┬  
-               │       ╰── Warning: Unused variable 'a'
-            ───╯
-            Finished. moon: ran 3 tasks, now up to date
-        "#]],
-    );
-
-    check(
-        get_err_stdout(&dir, ["build", "--deny-warn", "--sort-input"]),
-        expect![[r#"
-            failed: moonc build-package -error-format json -w @a-31-32 -alert @all-raise-throw-unsafe+deprecated $ROOT/lib/hello.mbt -o $ROOT/target/wasm-gc/release/build/lib/lib.core -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources username/hello/lib:$ROOT/lib -target wasm-gc
-        "#]],
-    );
-}
-
-#[test]
-fn test_moon_fmt() {
-    let dir = TestDir::new("moon_fmt.in");
-    check(
-        read(dir.join("lib").join("hello.mbt")),
-        expect![[r#"
-                pub fn hello() -> String { "Hello, world!" }
-            "#]],
-    );
-    check(
-        read(dir.join("main").join("main.mbt")),
-        expect![[r#"
-                fn main { println(@lib.hello()) }"#]],
-    );
-    let _ = get_stdout(&dir, ["fmt"]);
-    check(
-        read(dir.join("lib").join("hello.mbt")),
-        expect![[r#"
-            ///|
-            pub fn hello() -> String {
-              "Hello, world!"
-            }
-        "#]],
-    );
-    check(
-        read(dir.join("main").join("main.mbt")),
-        expect![[r#"
-            ///|
-            fn main {
-              println(@lib.hello())
-            }
-        "#]],
-    );
-}
-
-#[cfg(unix)]
-#[test]
-fn test_moon_fmt_002() {
-    let dir = TestDir::new("moon_fmt.in");
-    let _ = snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["fmt", "--check"])
-        .assert()
-        .failure()
-        .get_output()
-        .stdout
-        .to_owned();
-    check(
-        read(dir.join("lib").join("hello.mbt")),
-        expect![[r#"
-            pub fn hello() -> String { "Hello, world!" }
-        "#]],
-    );
-    check(
-        read(dir.join("main").join("main.mbt")),
-        expect![[r#"
-            fn main { println(@lib.hello()) }"#]],
-    );
-    check(
-        read(
-            dir.join("target")
-                .join(TargetBackend::default().to_dir_name())
-                .join("release")
-                .join("format")
-                .join("lib")
-                .join("hello.mbt"),
-        ),
-        expect![[r#"
-            ///|
-            pub fn hello() -> String {
-              "Hello, world!"
-            }
-        "#]],
-    );
-    check(
-        read(
-            dir.join("target")
-                .join(TargetBackend::default().to_dir_name())
-                .join("release")
-                .join("format")
-                .join("main")
-                .join("main.mbt"),
-        ),
-        expect![[r#"
-            ///|
-            fn main {
-              println(@lib.hello())
-            }
-        "#]],
-    );
-}
-
-#[test]
-fn test_moon_fmt_extra_args() {
-    let dir = TestDir::new("moon_fmt.in");
-    check(
-        get_stdout(&dir, ["fmt", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonfmt ./lib/hello.mbt -w -o ./target/wasm-gc/release/format/lib/hello.mbt
-            moonfmt ./lib/hello_wbtest.mbt -w -o ./target/wasm-gc/release/format/lib/hello_wbtest.mbt
-            moonfmt ./main/main.mbt -w -o ./target/wasm-gc/release/format/main/main.mbt
-        "#]],
-    );
-    check(
-        get_stdout(&dir, ["fmt", "--dry-run", "--sort-input", "--", "a", "b"]),
-        expect![[r#"
-            moonfmt ./lib/hello.mbt -w -o ./target/wasm-gc/release/format/lib/hello.mbt a b
-            moonfmt ./lib/hello_wbtest.mbt -w -o ./target/wasm-gc/release/format/lib/hello_wbtest.mbt a b
-            moonfmt ./main/main.mbt -w -o ./target/wasm-gc/release/format/main/main.mbt a b
-        "#]],
-    );
-    check(
-        get_stdout(&dir, ["fmt", "--check", "--sort-input", "--dry-run"]),
-        expect![[r#"
-            moon tool format-and-diff --old ./lib/hello.mbt --new ./target/wasm-gc/release/format/lib/hello.mbt
-            moon tool format-and-diff --old ./lib/hello_wbtest.mbt --new ./target/wasm-gc/release/format/lib/hello_wbtest.mbt
-            moon tool format-and-diff --old ./main/main.mbt --new ./target/wasm-gc/release/format/main/main.mbt
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            [
-                "fmt",
-                "--check",
-                "--sort-input",
-                "--dry-run",
-                "--",
-                "c",
-                "d",
-            ],
-        ),
-        expect![[r#"
-            moon tool format-and-diff --old ./lib/hello.mbt --new ./target/wasm-gc/release/format/lib/hello.mbt c d
-            moon tool format-and-diff --old ./lib/hello_wbtest.mbt --new ./target/wasm-gc/release/format/lib/hello_wbtest.mbt c d
-            moon tool format-and-diff --old ./main/main.mbt --new ./target/wasm-gc/release/format/main/main.mbt c d
-        "#]],
-    );
-}
-
-#[test]
-fn test_moon_fmt_block_style() {
-    let dir = TestDir::new("moon_fmt.in");
-    check(
-        get_stdout(&dir, ["fmt", "--block-style", "--sort-input", "--dry-run"]),
-        expect![[r#"
-            moonfmt ./lib/hello.mbt -w -o ./target/wasm-gc/release/format/lib/hello.mbt -block-style
-            moonfmt ./lib/hello_wbtest.mbt -w -o ./target/wasm-gc/release/format/lib/hello_wbtest.mbt -block-style
-            moonfmt ./main/main.mbt -w -o ./target/wasm-gc/release/format/main/main.mbt -block-style
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            ["fmt", "--block-style=true", "--sort-input", "--dry-run"],
-        ),
-        expect![[r#"
-            moonfmt ./lib/hello.mbt -w -o ./target/wasm-gc/release/format/lib/hello.mbt -block-style
-            moonfmt ./lib/hello_wbtest.mbt -w -o ./target/wasm-gc/release/format/lib/hello_wbtest.mbt -block-style
-            moonfmt ./main/main.mbt -w -o ./target/wasm-gc/release/format/main/main.mbt -block-style
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            ["fmt", "--block-style=false", "--sort-input", "--dry-run"],
-        ),
-        expect![[r#"
-            moonfmt ./lib/hello.mbt -w -o ./target/wasm-gc/release/format/lib/hello.mbt
-            moonfmt ./lib/hello_wbtest.mbt -w -o ./target/wasm-gc/release/format/lib/hello_wbtest.mbt
-            moonfmt ./main/main.mbt -w -o ./target/wasm-gc/release/format/main/main.mbt
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "fmt",
-                "--block-style",
-                "--check",
-                "--sort-input",
-                "--dry-run",
-            ],
-        ),
-        expect![[r#"
-            moon tool format-and-diff --old ./lib/hello.mbt --new ./target/wasm-gc/release/format/lib/hello.mbt --block-style
-            moon tool format-and-diff --old ./lib/hello_wbtest.mbt --new ./target/wasm-gc/release/format/lib/hello_wbtest.mbt --block-style
-            moon tool format-and-diff --old ./main/main.mbt --new ./target/wasm-gc/release/format/main/main.mbt --block-style
-        "#]],
-    );
 }
 
 #[test]
