+++ crates/moon/tests/test_cases/warns/alert_list/lib/moon.pkg.json
