+++ crates/moon/tests/test_cases/warns/alert_list/main/moon.pkg.json
