+++ crates/moon/tests/test_cases/warns/alert_list/moon.mod.json
