+++ crates/moon/tests/test_cases/warns/deny_warn/lib/hello.mbt
