+++ crates/moon/tests/test_cases/warns/deny_warn/lib/moon.pkg.json
