+++ crates/moon/tests/test_cases/warns/deny_warn/main/main.mbt
