+++ crates/moon/tests/test_cases/warns/deny_warn/main/moon.pkg.json
