+++ crates/moon/tests/test_cases/warns/deny_warn/moon.mod.json
