+++ crates/moon/tests/test_cases/warns/mod_level/lib/hello.mbt
