+++ crates/moon/tests/test_cases/warns/mod_level/lib/moon.pkg.json
