+++ crates/moon/tests/test_cases/warns/mod_level/main/moon.pkg.json
