+++ crates/moon/tests/test_cases/warns/warn_list/lib/moon.pkg.json
