+++ crates/moon/tests/test_cases/warns/warn_list/lib1/moon.pkg.json
