+++ crates/moon/tests/test_cases/warns/warn_list/main/moon.pkg.json
