+++ crates/moon/tests/test_cases/warns/warn_list/moon.mod.json
